+++ src/server.rs
@@ -3,6 +3,7 @@ use crate::console::CommandType;
 use crate::greet::greet;
 use crate::health::health_check;
 use crate::middleware;
+use crate::websocket;
 use axum::Router;
 use axum::routing::get;
 use std::sync::Arc;
@@ -19,6 +20,7 @@ pub async fn server(
     let app = Router::new()
         .route("/greet", get(greet))
         .route("/health", get(health_check))
+        .route("/websocket", get(websocket::handler::ws_handler))
         .nest("/api/v1", api_v1(state.clone()))
         .layer(axum::middleware::from_fn(middleware::tracing::tracing))
         .layer(axum::middleware::from_fn(middleware::error::panic))